Extract pose/keypoint visualization subsystem into engine/src/pose.ts
Moves normalizePose (parses Halpe/COCO/OpenPose/Human3.6M/generic pose
JSON variants), autoConnectKnn, buildPoseGroup, and the per-frame
update functions (appearance, labels, scaling, convention flip,
score/uncertainty filters) out of PointCloudVisualizer. normalizePose
and autoConnectKnn were already pure; the rest take an explicit
PoseHost parameter the same way sequencePlayback.ts does. main.ts
keeps thin delegating wrappers where call sites exist outside
handlePoseData; normalizePose/buildPoseGroup are called directly from
handlePoseData since they had no other callers.

Relaxed private on the ~13 fields/properties the module touches
(poseGroups, poseMeta, poseJoints, poseEdges, the per-pose toggle
arrays, fileColors) for the same structural-typing reason as the
sequence-playback extraction.

main.ts: 14,238 -> 13,625 lines.

Verified: engine tsc clean, full webpack compile clean, all 59
extension tests passing under xvfb-run.
